refactor: migrate log commands to package
